+++ src/MimesisPlayerEnhancement/Features/UserInterface/LoadingWaitPlayerList/LoadingWaitPlayerListPatches.cs
@@ -1,0 +1,6 @@
+namespace MimesisPlayerEnhancement.Features.UserInterface.LoadingWaitPlayerList
+{
+    internal static class LoadingWaitPlayerListPatches
+    {
+    }
+}